farmacia/requirements.txt
python-dateutil
request